[SD] Fix multiple call to device check
- Also makes the dark theme default.
- Fix custom_vae parameter in img2img.
