+++ apps/stable_diffusion/scripts/inpaint.py
@@ -220,6 +220,7 @@ def inpaint_inf(
             args.import_mlir,
             args.hf_model_id,
             args.ckpt_loc,
+            args.custom_vae,
             args.precision,
             args.max_length,
             args.batch_size,
@@ -301,6 +302,7 @@ if __name__ == "__main__":
         args.import_mlir,
         args.hf_model_id,
         args.ckpt_loc,
+        args.custom_vae,
         args.precision,
         args.max_length,
         args.batch_size,
